Key1 was easy money, key2 not so much..
I was actually able to do key2 if the numbers were justified either fully to the left or fully to the right. Turns out they weren't, I tried padding them with zeros but even then it was giving me trouble because of the way I had splitted and stripped the dataset. Once I discovered this, I was in shambles.